% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_resources.R
\name{all_resources}
\alias{all_resources}
\title{Get Full Cluster Rundown}
\usage{
all_resources()
}
\value{
\code{data.frame} of values for each person
}
\description{
Get output of resources and slots begin used by cluster
}
