% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_tasks.R
\name{get_error_tasks}
\alias{get_error_tasks}
\title{Get Tasks in Error}
\usage{
get_error_tasks()
}
\value{
If no tasks, \code{\link{NULL}}, otherwise a 
\code{data.frame} of the tasks to be removed
}
\description{
Get Tasks in Error
}
\examples{
\dontrun{
get_error_tasks()
}
}
