% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_info.R
\name{job_info}
\alias{job_info}
\title{Get Job Information on Cluster}
\usage{
job_info(job_id, last_days = 100, verbose = TRUE)
}
\arguments{
\item{job_id}{Job identifier}

\item{last_days}{jobs started during the last \code{last_days} days,
passed to \code{qacct -d} argument}

\item{verbose}{print diagnostic messages}
}
\value{
A \code{tibble} of information
}
\description{
Get Job Information on Cluster
}
\examples{
\dontrun{
job_info("6684455")
}
}
