% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundown.R
\name{unmask.bandit}
\alias{unmask.bandit}
\title{Match alias to name and email}
\usage{
unmask.bandit(username)
}
\arguments{
\item{username}{Used if you want to see a specific user usage}
}
\value{
vector of stuff
}
\description{
Converts alias to name and email
}
