% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{full.rundown}
\alias{full.rundown}
\title{Get Full Cluster Rundown}
\usage{
full.rundown(blameGame = FALSE)
}
\arguments{
\item{blameGame}{logical, should return real names and emails?}
}
\value{
List of stuff
}
\description{
Get output of resources and slots begin used by cluster
}
