% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundown.R
\name{get.resource}
\alias{get.resource}
\title{Get list of node resources/memory}
\usage{
get.resource(username = NULL)
}
\arguments{
\item{username}{Used if you want to see a specific user usage}
}
\value{
List of stuff
}
\description{
Grabs the rundown of each compute node memory and slots
}
